Refactoring file handling a bit

+++ App/src/file/FileDataType.h
@@ -6,7 +6,7 @@ namespace med
 	{
 		Undefined,
 		Uint16,
-		Float32,
+		Uint32,
 		Double
 	};
 }
